Deleted the main scene to fix the issues we're having
Main scene got destroyed somehow and we can't figure out a way to restore it, so I deleted it for now in hopes we can either remake or find an older scene to replace it.  Also tidied up the some errors that I guess occured when Blake pushed his stuff.  Why....